first big refactor
