+++ setup/images.py
@@ -4,17 +4,28 @@ import pygame
 
 class Images:
 
+    BACKGROUND_IMG = pygame.image.load(os.path.join('media', 'board.png'))
+
     def __init__(self):
-        # Set up background and sizes
-        self.background = pygame.image.load(os.path.join('media', 'board.png')).convert()
-        self.pieces_image = pygame.image.load(os.path.join('media', 'Chess_Pieces_Sprite.png')).convert_alpha()
-        self.size_of_bg = self.background.get_rect().size
+        """Make a window of the same size as the background, set its title, and
+        load the background image onto it (the board)"""
+
+        # Background
+        self.size_of_bg = self.BACKGROUND_IMG.get_rect().size
+        self.screen = pygame.display.set_mode(self.size_of_bg)
+        self.background = self.BACKGROUND_IMG.convert()
+        self.screen.blit(self.background, (0, 0))
+
         self.square_width = self.size_of_bg[0] // 8
         self.square_height = self.size_of_bg[1] // 8
 
-        # Set up piece images
-        self.images = self.get_orig_images()
+        pygame.display.set_caption('Chess')
 
+        # Pieces
+        self.pieces_image = pygame.image.load(os.path.join('media', 'Chess_Pieces_Sprite.png')).convert_alpha()
+        self.pieces_image = pygame.transform.scale(self.pieces_image, (self.square_width * 6, self.square_height * 2))
+        self.images = self.get_orig_images()
+      
     def get_orig_images(self):
         """Load all the media into one dictionary"""
         circle_image_green = pygame.image.load(os.path.join('media', 'green_circle_small.png')).convert_alpha()
